♻️ [refactor]: event 클래스 삭제 및 dto 리팩토링
