+++ src/main/java/com/PickOne/global/verification/dto/EmailMessage.java
@@ -1,0 +1,8 @@
+package com.PickOne.global.verification.dto;
+
+public record EmailMessage(
+        String to,
+        String subject,
+        String body,
+        boolean isHtml
+) {}
