--- src/main/java/com/PickOne/global/verification/dto/EmailRequest.java
@@ -1,10 +1,0 @@
-package com.PickOne.global.verification.dto;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-
-public record EmailRequest(
-        @NotBlank(message = "이메일은 필수입니다.")
-        @Email(message = "유효한 이메일 형식이어야 합니다.")
-        String email
-) {}
