--- src/main/java/com/PickOne/global/verification/dto/TokenRequest.java
@@ -1,8 +1,0 @@
-package com.PickOne.global.verification.dto;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record TokenRequest(
-        @NotBlank(message = "토큰은 필수입니다.")
-        String token
-) {}
