+++ src/main/java/com/PickOne/global/verification/dto/TokenVerificationRequestDto.java
@@ -1,0 +1,5 @@
+package com.PickOne.global.verification.dto;
+
+public record TokenVerificationRequestDto(
+        String token
+) {}
