--- src/main/java/com/PickOne/global/verification/event/EmailVerifiedEvent.java
@@ -1,17 +1,0 @@
-package com.PickOne.global.verification.event;
-
-import lombok.Getter;
-import org.springframework.context.ApplicationEvent;
-
-@Getter
-public class EmailVerifiedEvent extends ApplicationEvent {
-
-    private final Long userId;
-    private final String email;
-
-    public EmailVerifiedEvent(Object source, Long userId, String email) {
-        super(source);
-        this.userId = userId;
-        this.email = email;
-    }
-}
